[Refactor] Separate HapticManager and Remove comments
