+++ Chinggu/Chinggu/Utils/HapticManager.swift
@@ -1,0 +1,22 @@
+//
+//  HapticManager.swift
+//  Chinggu
+//
+//  Created by Sebin Kwon on 2023/08/31.
+//
+
+import SwiftUI
+
+class HapticManager {
+    static let instance = HapticManager()
+    
+    func notification(type: UINotificationFeedbackGenerator.FeedbackType) {
+        let generator = UINotificationFeedbackGenerator()
+        generator.notificationOccurred(type)
+    }
+    
+    func impact(style: UIImpactFeedbackGenerator.FeedbackStyle) {
+        let generator = UIImpactFeedbackGenerator(style: style)
+        generator.impactOccurred()
+    }
+}
